fixed type error

+++ skimage/filters/window.py
@@ -70,7 +70,7 @@ def get_window(window, size, ndim=2, **kwargs):
     w = np.reshape(w, (-1,) + (1,) * (ndim-1))
 
     # Create coords for warping following `ndimage.map_coordinates` convention.
-    L = [np.arange(size) for i in range(ndim)]
+    L = [np.arange(size, dtype=np.double) for i in range(ndim)]
     coords = np.stack((np.meshgrid(*L)))
     center = (size / 2) - 0.5
     coords[0, ...] = np.sqrt(((coords - center) ** 2).sum(axis=0))
